Fix color in data struct

+++ wallbrick_struct_buffer.py
@@ -46,7 +46,7 @@ struct data
     int pos_y;
     int width;
     int height;
-    float4 color;
+    uint4 color;
 };
 StructuredBuffer<data> quads : register(t0);
 RWTexture2D<float4> target : register(u0);
